+++ alphatwirl/progressbar/ProgressReporter.py
@@ -26,17 +26,17 @@ class ProgressReporter(object):
     This method can be frequently called multiple times. However,
     after sending one report, the reporter wait for a certain
     ``interval`` (0.1 seconds by default) before sending another
-    report. Reports received within this interval will be discarded.
-    The exception for this is the last report. The last report, which
-    indicates the completion of the task, will be always sent to the
-    progress monitor regardless of whether it is given within the
-    interval.
+    report of the same task. Reports from the same task received
+    within this interval will be discarded. The exception for this is
+    the last report. The last report, which indicates the completion
+    of the task, will be always sent to the progress monitor
+    regardless of whether it is given within the interval.
 
     """
     def __init__(self, queue):
         self.queue = queue
         self.interval = 0.1 # [second]
-        self._read_time()
+        self.last_time = { } # key: taskid
 
     def __repr__(self):
         return '{}(queue={!r}, interval={!r})'.format(
@@ -53,21 +53,30 @@ class ProgressReporter(object):
 
         """
 
-        if not self._need_to_report(report): return
-        self._report(report)
+        if not self._need_to_report(report):
+            return
 
-    def _report(self, report):
         self.queue.put(report)
-        self._read_time()
+
+        self.last_time[report.taskid] = self._time()
 
     def _need_to_report(self, report):
-        if self._time() - self.last_time > self.interval: return True
-        if report.last(): return True
-        if report.first(): return True
+
+        if report.first():
+            return True
+
+        if report.last():
+            return True
+
+        if report.taskid not in self.last_time:
+            return True
+
+        if self._time() - self.last_time[report.taskid] > self.interval:
+            return True
+
         return False
 
-    def _read_time(self): self.last_time = self._time()
-
-    def _time(self): return time.time()
+    def _time(self):
+        return time.time()
 
 ##__________________________________________________________________||